g2 图表模板完成
