+++ scaffold/package.json
@@ -10,7 +10,7 @@
   "dependencies": {
     "antd": "^2.12.1",
     "dva": "^1.2.1",
-    "g2": "^2.3.7-beta.3",
+    "g2": "^2.3.8",
     "g2-plugin-slider": "^1.2.1",
     "marked": "^0.3.6",
     "numeral": "^2.0.6",
