+++ scaffold/src/components/Charts/NumberInfo/index.less
@@ -37,3 +37,10 @@
     }
   }
 }
+.numberInfolight {
+  & > div {
+    & > span {
+      color: @text-color;
+    }
+  }
+}
