+++ scaffold/src/routes/Dashboard/Analysis.less
@@ -1,5 +1,6 @@
-.normal {
-  font-family: Georgia, sans-serif;
-  margin-top: 3em;
-  text-align: center;
+.iconGroup {
+  i {
+    cursor: pointer;
+    margin-left: 16px;
+  }
 }
