+++ scaffold/src/routes/Dashboard/Monitor.less
@@ -1,6 +1,5 @@
-.iconGroup {
-  i {
-    cursor: pointer;
-    margin-left: 16px;
-  }
+.normal {
+  font-family: Georgia, sans-serif;
+  margin-top: 3em;
+  text-align: center;
 }
